[test] Re-organized tests to fit the new API
